+++ src/encoders/shn.c
@@ -495,7 +495,7 @@ ShortenEncoder_put_long(Bitstream* bs, int value)
 }
 
 void
-ShortenEncoder_byte_counter(int byte, void* counter)
+ShortenEncoder_byte_counter(uint8_t byte, void* counter)
 {
     int* i_counter = (int*)counter;
     *i_counter += 1;
